Resolved dropbox conflicts
