+++ PrototypeMenu.cpp
@@ -1,46 +1,71 @@
 #include "PrototypeMenu.h"
 
-MenuBox::MenuBox()
-{
 
+
+Menu::Menu()
+{
+	// Change to reletive coordinants when available
+	int x = 0;
+	int y = 0;
+
+	int xSize = 200;
+	int ySize = 400;
+
+	int xPos = (x / 2) - (xSize / 2);
+	int yPos = (y / 2) - (ySize / 2);
+
+
+	// Create a menu box
+	// Position is set centered in the view
+	menuBox.setPosition(xPos, yPos);
+	menuBox.setSize(sf::Vector2f(xSize, ySize));
+	menuBox.setFillColor(sf::Color(0, 0, 0, 220));
+	menuBox.setOutlineThickness(3.f);
+	menuBox.setOutlineColor(sf::Color(100, 149, 237));
 }
 
 
 
-void MenuBox::update(sf::Time dt)
+void Menu::update(sf::Time dt)
 {
-
+  // Check for input to:
+  // 1. Change cursor position
+  // 2. Select and option
 }
 
 
 
-void MenuBox::draw(sf::RenderTarget& target, sf::RenderStates state) const
+void Menu::draw(sf::RenderWindow& window)
 {
-
+	window.draw(menuBox);	
+  // draw options
+  // draw cursor position
 }
 
 
+
+StartMenu::StartMenu()
+{
+  //Option * start = new OptStart(); 
+  //Option * quit = new OptQuit(); 
+  options.push_back(std::unique_ptr<Option>(new OptStart));
+  options.push_back(std::unique_ptr<Option>(new OptQuit));
+}
+
+
+
+PauseMenu::PauseMenu()
+{
+  //Option * cont = new OptCont(); 
+  //Option * quit = new OptQuit(); 
+  options.push_back(std::unique_ptr<Option>(new OptCont));
+  options.push_back(std::unique_ptr<Option>(new OptQuit));
+}
+
+
+
+// Does this need to be here?
 Option::Option()
-{
-
-}
-
-
-
-PrototypeMenu::PrototypeMenu()
-{
-
-}
-
-
-
-void PrototypeMenu::update(sf::Time dt)
-{
-
-}
-
-
-void PrototypeMenu::draw(sf::RenderWindow& window)
 {
 
 }
@@ -49,7 +74,20 @@ void PrototypeMenu::draw(sf::RenderWindow& window)
 
 OptStart::OptStart()
 {
+  if (!menuFont.loadFromFile("AreaKilometer50.otf")) {
+    // error
+  }
+  lable.setFont(menuFont);
+  lable.setString("Start");
+  lable.setCharacterSize(24);
+  lable.setFillColor(sf::Color::White);
+}
 
+
+
+void OptStart::draw(sf::RenderWindow& window)
+{
+  window.draw(lable);
 }
 
 
@@ -68,6 +106,7 @@ bool OptStart::onSelection()
 
 OptCont::OptCont()
 {
+  // Set the text object
 
 }
 
@@ -88,6 +127,7 @@ bool OptCont::onSelection()
 
 OptQuit::OptQuit()
 {
+  // Set the text object
 
 }
 
@@ -95,9 +135,11 @@ OptQuit::OptQuit()
 
 bool OptQuit::onSelection()
 {
+  // Pop all of the scenes off of the stack
   while (System.popScene()) ;
 
-  // System.quit()?
+  // Quit the program
+  System.quit();
 
   return true;
 }
